add the correct_rate of predictor && optimize the variable types

+++ head_file/register_file.h
@@ -1,6 +1,8 @@
 #ifndef RISC_V_SIMULATOR_REGISTER_FILE_H
 #define RISC_V_SIMULATOR_REGISTER_FILE_H
 
+#include "tool.h"
+
 class register_file {
 private:
 
@@ -15,16 +17,16 @@ public:
     void init();
 
     //获得寄存器的依赖关系和值
-    void get(int reg_, int &depend_, int &value);
+    void get(u_int8_t reg_, int &depend_, int &value);
 
     //获得目标寄存器的值（仅用于store的commit，不检查依赖关系）
-    int get(int reg_);
+    int get(u_int8_t reg_);
 
     //指令完成提交后，删除相关标记，并更新其值
-    void flush_tag(int tag_, int data_, int reg_);
+    void flush_tag(int tag_, int data_, u_int8_t reg_);
 
     //给reg_寄存器存标记
-    void set_tag(int tag_, int reg_);
+    void set_tag(int tag_, u_int8_t reg_);
 
     //得到a10的返回值
     int get_return_value();
